tests/wpt/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-cancel-preventDefault-closewatcher.tentative.html

<!DOCTYPE html>
<html>
<head>
  <title>Test cancel event with preventDefault on cancel event for dialog element</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <link rel="help" href="https://bugs.webkit.org/show_bug.cgi?id=227534">
  <link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1322947">
</head>
<body>
<p>Test cancel event with preventDefault on cancel event for dialog element</p>
<dialog>
  <p>Hello World</p>
  <button>user activation button</button>
</dialog>
<script>
  setup({ single_test: true });

  var hasCancelEventFired = false;

  const dialog = document.querySelector("dialog");

  const verify = () => {
    assert_true(hasCancelEventFired, "cancel is fired");
    done();
  };

  dialog.addEventListener("cancel", function(event) {
    hasCancelEventFired = true;
    event.preventDefault();
    step_timeout(function() {
      verify();
    }, 0)
  });

  dialog.addEventListener("close", function() {
    assert_true(false, "close event should not be fired");
  });

  dialog.showModal();

  (async () => {
    // Pressing escape on the dialog needs user activation or else the cancel event won't be fired.
    const button = dialog.querySelector('button');
    const buttonClickPromise = new Promise(resolve => button.onclick = resolve);
    await test_driver.click(button);
    await buttonClickPromise;

    test_driver.send_keys(document.documentElement, "\uE00C"); // ESC key
  })();
</script>
</body>
</html>
